+++ gherkin/dotnet/Gherkin/Events/Args/Ast/Comment.cs
@@ -1,0 +1,12 @@
+﻿using System.Runtime.Serialization;
+
+namespace Gherkin.Events.Args.Ast
+{
+    public class Comment
+    {
+        [DataMember(Name="text")]
+        public string Text { get; set; }
+        [DataMember(Name = "location")]
+        public Location Location { get; set; }
+    }
+}
