+++ gherkin/dotnet/Gherkin/Events/GherkinDocumentEvent.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.Serialization;
 using Gherkin.Events.Args;
+using Gherkin.Events.Args.Ast;
 
 namespace Gherkin.Events
 {
